// src/utility/hero-types.ts
export type HeroName =
  | "abaddon"
  | "alchemist"
  | "ancient apparition"
  | "anti-mage"
  | "arc warden"
  | "axe"
  | "bane"
  | "batrider"
  | "beastmaster"
  | "bloodseeker"
  | "bounty hunter"
  | "brewmaster"
  | "bristleback"
  | "broodmother"
  | "centaur warrunner"
  | "chaos knight"
  | "chen"
  | "clinkz"
  | "clockwerk"
  | "crystal maiden"
  | "dark seer"
  | "dark willow"
  | "dawnbreaker"
  | "dazzle"
  | "death prophet"
  | "disruptor"
  | "doom"
  | "dragon knight"
  | "drow ranger"
  | "earth spirit"
  | "earthshaker"
  | "elder titan"
  | "ember spirit"
  | "enchantress"
  | "enigma"
  | "faceless void"
  | "grimstroke"
  | "gyrocopter"
  | "hoodwink"
  | "huskar"
  | "invoker"
  | "io"
  | "jakiro"
  | "juggernaut"
  | "keeper of the light"
  | "kunkka"
  | "legion commander"
  | "leshrac"
  | "lich"
  | "lifestealer"
  | "lina"
  | "lion"
  | "lone druid"
  | "luna"
  | "lycan"
  | "magnus"
  | "marci"
  | "mars"
  | "medusa"
  | "meepo"
  | "mirana"
  | "monkey king"
  | "muerta"
  | "morphling"
  | "nagasiren"
  | "furion"
  | "necrophos"
  | "night stalker"
  | "nyx assassin"
  | "ogre magi"
  | "omniknight"
  | "oracle"
  | "outworld destroyer"
  | "pangolier"
  | "phantom assassin"
  | "phantom lancer"
  | "phoenix"
  | "puck"
  | "pudge"
  | "pugna"
  | "primal beast"
  | "queen of pain"
  | "razor"
  | "riki"
  | "rubick"
  | "sand king"
  | "shadow demon"
  | "shadow fiend"
  | "shadow shaman"
  | "silencer"
  | "skywrath mage"
  | "slardar"
  | "slark"
  | "snap fire"
  | "sniper"
  | "spectre"
  | "spirit breaker"
  | "storm spirit"
  | "sven"
  | "techies"
  | "templar assassin"
  | "terrorblade"
  | "tidehunter"
  | "timbersaw"
  | "tinker"
  | "tiny"
  | "treant protector"
  | "troll warlord"
  | "tusk"
  | "underlord"
  | "undying"
  | "ursa"
  | "vengeful spirit"
  | "venomancer"
  | "viper"
  | "visage"
  | "void spirit"
  | "warlock"
  | "weaver"
  | "windranger"
  | "winter wyvern"
  | "witch doctor"
  | "wraith king"
  | "zeus";

export type Complexity = "1" | "2" | "3";

export type Attribute = "agility" | "strength" | "intelligence" | "universal";

export type Dota2Hero = {
  name: HeroName;
  image: string;
  attribute: Attribute;
  complexity: Complexity;
};

export type Dota2Heroes = Array<Dota2Hero>